move pages to lib
