+++ apps/web/middleware.ts
@@ -1,6 +1,6 @@
 import { withAuth } from "next-auth/middleware";
 
-import { pages } from "./server/pages";
+import { pages } from "./lib/pages";
 
 export default withAuth({
   pages,
